feat(io): size the process image per board, and refuse locations outside it
Every arduino-cli target sized its I/O image from one hardcoded set of
MAX_* macros in openplc.h, so a P1AM with a 15-slot expansion backplane
(up to 240 discrete points) got the same 56 an Arduino Uno gets. The
reporter hit the ceiling at %QX7.0 (openplc-editor#296).

The ceiling was not the whole problem. Nothing checked it either:

- runtime_bind_located_vars() wrote bool_input/bool_output/int_*/​*_memory
  at whatever byte_index the descriptor carried. Only the DWord cases had
  a bound. %QX7.0 on a 56-output image indexes bool_output[7][8] -- one
  past the end -- and corrupted whatever followed it.
- The editor happily allocated and compiled such an address. The Python
  editor refused it at glue-code generation ("wrong location for var
  __QX7_0"); that check did not survive the move to strucpp, so between
  then and now the I/O silently did nothing.
- init_mbregs() took its sizes as uint8_t, so 256 coils became 0, and
  readCoils() truncated the coil address to uint8_t, aliasing every coil
  above 255 onto a low one and answering with the wrong bit.

So: make the sizes a per-target capability sourced from the VPP manifest
(riding the path isLicensable already uses), emit them into defines.h,
and have openplc.h take them through #ifndef guards. A target that
declares none emits nothing and keeps compiling on the header's own
#ifdef ladder -- byte-for-byte its current defines.h, which the emitter's
snapshot tests pin.

Then close the gaps: bound every slot write in the glue, widen the Modbus
sizes and the coil address to 16 bits, and add a validate step that
refuses an out-of-range location before the build starts, naming the
board and the limit.

Also replace mapEmptyBuffers()'s malloc(1)-per-unbound-point with one
static block. At 240+240 points that was ~480 one-byte allocations, each
with its own heap header, fragmenting the heap before the first scan.

Editor side of openplc-editor#565 rides along, because the two share this
code path: a located ARRAY claims one slot per element, so the range
check measures its last element, and location validation checks the
ELEMENT type against the address class instead of rejecting arrays
outright. The empty error message for an unlocatable type is fixed too --
it rendered as a refusal with no reason at all.

+++ resources/sources/Baremetal/modbus_registers.cpp
@@ -9,7 +9,7 @@ Copyright (C) 2022 OpenPLC - Thiago Alves
 // In a debug-only build this whole TU compiles to nothing, saving flash/SRAM.
 #ifdef MODBUS_ENABLED
 
-bool init_mbregs(uint8_t size_holding, uint8_t size_dint_memory, uint8_t size_lint_memory, uint8_t size_coils, uint8_t size_inputregs, uint8_t size_inputstatus)
+bool init_mbregs(uint16_t size_holding, uint16_t size_dint_memory, uint16_t size_lint_memory, uint16_t size_coils, uint16_t size_inputregs, uint16_t size_inputstatus)
 {
     //Save sizes
     modbus.holding_size = size_holding;
@@ -62,9 +62,14 @@ bool init_mbregs(uint8_t size_holding, uint8_t size_dint_memory, uint8_t size_li
     return true;
 }
 
+// byte_addr is uint16_t, not uint8_t: addr is already a 16-bit Modbus
+// address, so addr/8 overflows a uint8_t past 2040 coils. The callers
+// bound `addr` against *_size before getting here, and *_size is now
+// itself 16-bit -- narrowing the index would put the truncation back one
+// step further down.
 bool get_discrete(uint16_t addr, bool regtype)
 {
-    uint8_t byte_addr = addr / 8;
+    uint16_t byte_addr = addr / 8;
     uint8_t bit_addr = addr % 8;
     if (regtype == COILS)
         return bitRead(modbus.coils[byte_addr], bit_addr);
@@ -74,7 +79,7 @@ bool get_discrete(uint16_t addr, bool regtype)
 
 void write_discrete(uint16_t addr, bool regtype, bool value)
 {
-    uint8_t byte_addr = addr / 8;
+    uint16_t byte_addr = addr / 8;
     uint8_t bit_addr = addr % 8;
     if (regtype == COILS)
         bitWrite(modbus.coils[byte_addr], bit_addr, value);
@@ -116,7 +121,9 @@ void readRegisters(uint16_t startreg, uint16_t numregs)
 
     uint16_t val;
     uint16_t i = 0;
-    uint8_t pos = 0;
+    // uint16_t, not uint8_t: pos indexes dint_memory/lint_memory, whose
+    // sizes come from the MAX_MEMORY_* macros and are no longer capped at 255.
+    uint16_t pos = 0;
 	while(numregs--)
     {
         if ((startreg + i) < modbus.holding_size)
@@ -177,7 +184,9 @@ void writeSingleRegister(uint16_t reg, uint16_t value)
         return;
     }
 
-    uint8_t pos = 0;
+    // uint16_t, not uint8_t: pos indexes dint_memory/lint_memory, whose
+    // sizes come from the MAX_MEMORY_* macros and are no longer capped at 255.
+    uint16_t pos = 0;
 
     if (reg < modbus.holding_size)
     {
@@ -254,7 +263,9 @@ void writeMultipleRegisters(uint16_t startreg, uint16_t numoutputs, uint8_t byte
 
     uint16_t value;
     uint16_t i = 0;
-    uint8_t pos = 0;
+    // uint16_t, not uint8_t: pos indexes dint_memory/lint_memory, whose
+    // sizes come from the MAX_MEMORY_* macros and are no longer capped at 255.
+    uint16_t pos = 0;
 	while(numoutputs--)
     {
         value = (uint16_t)mb_frame[7+i*2] << 8 | (uint16_t)mb_frame[8+i*2];
@@ -350,7 +361,12 @@ void readCoils(uint16_t startreg, uint16_t numregs)
 	while (numregs)
     {
         i = (totregs - numregs--) / 8;
-		if (get_discrete((uint8_t)startreg, COILS))
+		// No (uint8_t) cast on startreg: it is a 16-bit coil address, and
+		// truncating it aliased every coil above 255 onto a low one --
+		// FC 0x01 answered with the wrong bit and no error. Harmless while
+		// no board had more than 56 coils; reachable as soon as one does
+		// (openplc-editor#296). readInputStatus below never had the cast.
+		if (get_discrete(startreg, COILS))
 			bitSet(mb_frame[3+i], bitn);
 		else
 			bitClear(mb_frame[3+i], bitn);
